pushing so Daniela can help me figure out how to pass distance from one component to another
