+++ that-guide/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "T.H.A.T. Guide",
+  "name": "Thru Hiker's Appalachian Trail Guide",
   "icons": [
     {
       "src": "hiker.png",
