--- that-guide/src/App.css
@@ -1,38 +1,0 @@
-/* .title-header {
-  border: 2px solid black;
-  text-align: center;
-  line-height: .6;
-}
-
-.nav-bar {
-  text-align: center;
-}
-
-.location-header {
-  text-align: center;
-  border: 2px solid black;
-  line-height: .6;
-  width: 60%;
-  margin: auto;
-}
-
-.map-container {
-  border: 4px solid black;
-  margin-top: 10px;
-  height: 400px;
-  width: 200px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-
-.hike-starter {
-  border: 2px solid black;
-  text-align: center;
-  margin-top: 20px;
-}
-
-.current-hike-stats {
-  border: 2px solid black;
-  text-align: center;
-} */
