+++ that-guide/src/App.js
@@ -1,8 +1,8 @@
-import "./App.css";
+
 import { useEffect, useState } from "react";
 import axios from "axios";
 import Homepage from "./components/homepage/Homepage";
-import StartHike from "./components/StartHike";
+import StartHike from "./components/StartHike/StartHike";
 import {Routes, Route, BrowserRouter as Router} from "react-router-dom"
 import useLocalStorageState from "use-local-storage-state";
 
@@ -11,7 +11,7 @@ function App() {
   const [description, setDescription] = useState("");
   const [memeImage, setMemeImage] = useState("");
   const [team, setTeam] = useState("");
-  const [distance, setDistance] = useState(null)
+  const [distance, setDistance] = useState("5")
 
   return (
     <>
