+++ that-guide/src/components/StartHike/StartHike.css
@@ -1,0 +1,14 @@
+.whole-stats-container {
+    display: flex;
+    justify-content: space-between;
+}
+
+.left-container {
+    text-align: left;
+    padding: 10px;
+}
+
+.right-container {
+    text-align: right;
+    padding: 10px;
+}
