+++ that-guide/src/components/homepage/Homepage.js
@@ -45,11 +45,10 @@ export default function Homepage() {
 
   return (
     <>
-      <div className="location-header">
+      {/* <div className="location-header">
         <h3>Current Location: {latitude}, {longitude}
-          {/* <CurrentLocation lat={lat} long={long} /> */}
         </h3>
-      </div>
+      </div> */}
 
       <div className="map-and-button"> 
         <div className="homepage-map">
@@ -60,24 +59,15 @@ export default function Homepage() {
 
       <div className="hike-starter">
         <div className="current-stats">
-          <h3>Current Coordinates: (list coordinates)</h3>
+          <h3>Current Coordinates: {latitude}, {longitude}</h3>
           <h3>Current Elevation: (list elevation)</h3>
         </div>
         <h2>How far do you want to hike?</h2>
-        <form id="select-distance">Select Distance:</form>
-        <select value={distance} onChange={handleSetDistance}>
-          <option value="">Please Select a Distance</option>
-          <option value="1">1 mile</option>
-          <option value="2">2 miles</option>
-          <option value="3">3 miles</option>
-          <option value="4">4 miles</option>
-          <option value="5">5 miles</option>
-          <option value="6">6 miles</option>
-          <option value="7">7 miles</option>
-          <option value="8">8 miles</option>
-          <option value="9">9 miles</option>
-          <option value="10">10 miles</option>
-        </select>
+        <form id="select-distance" onSubmit={setDistance}>Select Distance (in miles):</form>
+          <input 
+            type="number"
+            value={distance}
+            onChange={(e) => setDistance(e.target.value)} />
         <button type="submit" className="start-hike" onClick={handleStartHike} onSubmit={setDistance}>
           Start your hike!
         </button>
