+++ that-guide/src/components/homepage/homepage.css
@@ -20,7 +20,7 @@
   border: 4px solid black;
   margin-top: 10px;
   height: 400px;
-  width: 200px;
+  width: 300px;
   margin-left: auto;
   margin-right: auto;
 }
